requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.5
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
rawpy==0.14.0
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
tifffile==2020.10.1
